instantButChangeable meeting stuff

+++ Games/types/LiarsDice/items/Microphone.js
@@ -11,7 +11,7 @@ module.exports = class Microphone extends Item {
       "Amount": {
         actionName: "How many?",
         states: ["Guess Dice"],
-        flags: ["voting", "instantAction"],
+        flags: ["voting", "instant", "instantButChangeable", "repeatable"],
         inputType: "text",
         textOptions: {
           minLength: 1,
@@ -22,36 +22,68 @@ module.exports = class Microphone extends Item {
         action: {
           item: this,
           run: function () {
+            this.actor.howManySelected = true;
             this.actor.role.data.amount = this.target;
+
+            if (this.actor.whichFaceSelected) {
+
+              this.game.lastBidder = this.actor;
+              this.game.lastAmountBid = this.actor.role.data.amount;
+              this.game.lastFaceBid = this.target;
+              this.game.sendAlert(`${this.actor.name} guesses ${this.actor.role.data.amount}x ${this.actor.role.data.face}`);
+              this.item.drop();
+
+              this.actor.getMeetings().forEach((meeting) => {
+                if (meeting.name == "Amount" || meeting.name == "Face" || meeting.name == "CallLie" || meeting.name == "SpotOn") {
+                  meeting.leave(this.actor, true);
+                }
+              });
+
+            }
           },
         },
       },
       "Face": {
         actionName: "Which face?",
         states: ["Guess Dice"],
-        flags: ["voting"],
+        flags: ["voting", "instant", "instantButChangeable", "repeatable"],
         inputType: "custom",
         targets: ["1", "2", "3", "4", "5", "6"],
+        canUnvote: false,
         action: {
           item: this,
           run: function () {
-            this.game.lastBidder = this.actor;
-            this.game.lastAmountBid = this.actor.role.data.amount;
-            this.game.lastFaceBid = this.target;
-            this.game.sendAlert(`${this.actor.name} guesses ${this.actor.role.data.amount}x ${this.target}`);
-            this.item.drop();
+            this.actor.whichFaceSelected = true;
+            this.actor.role.data.face = this.target;
+
+            if (this.actor.howManySelected) {
+
+              this.game.lastBidder = this.actor;
+              this.game.lastAmountBid = this.actor.role.data.amount;
+              this.game.lastFaceBid = this.target;
+              this.game.sendAlert(`${this.actor.name} guesses ${this.actor.role.data.amount}x ${this.actor.role.data.face}`);
+              this.item.drop();
+
+              this.actor.getMeetings().forEach((meeting) => {
+                if (meeting.name == "Amount" || meeting.name == "Face" || meeting.name == "CallLie" || meeting.name == "SpotOn") {
+                  meeting.leave(this.actor, true);
+                }
+              });
+
+            }
           },
         },
       },
       "CallLie": {
-        actionName: "Call A Lie!",
+        actionName: "Call A Lie?",
         states: ["Guess Dice"],
         flags: ["voting", "instant"],
-        inputType: "boolean",
+        inputType: "button",
+        targets: ["Yes!"], 
         action: {
           item: this,
           run: function () {
-            if (this.target == "Yes") {
+            if (this.target == "Yes!") {
               this.actor.getMeetings().forEach((meeting) => {
                 if (meeting.name == "Amount" || meeting.name == "Face" || meeting.name == "CallLie" || meeting.name == "SpotOn") {
                   meeting.leave(this.actor, true);
@@ -69,14 +101,15 @@ module.exports = class Microphone extends Item {
 
     if (this.game.spotOn) {
       this.meetings["SpotOn"] = {
-        actionName: "Spot On!",
+        actionName: "Call a Spot On?",
         states: ["Guess Dice"],
         flags: ["voting", "instant"],
-        inputType: "boolean",
+        inputType: "button",
+        targets: ["Yes!"], 
         action: {
           item: this,
           run: function () {
-            if (this.target == "Yes") {
+            if (this.target == "Yes!") {
 
               if (this.game.lastBidder != null) {
                 this.actor.getMeetings().forEach((meeting) => {
@@ -98,6 +131,9 @@ module.exports = class Microphone extends Item {
   hold(player) {
     super.hold(player);
     player.game.queueAlert(`${player.name} is guessing dice…`);
+
+    player.howManySelected = false;
+    player.whichFaceSelected = false;
 
     this.setupMeetings();
